Move extension method down
