+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Extensions/StringExtensions.cs
@@ -1,10 +1,23 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Microsoft.AspNetCore.Razor;
+
 namespace Microsoft.CodeAnalysis.Razor;
 
 internal static class StringExtensions
 {
+    private const string RazorExtension = ".razor";
+    private const string CSHtmlExtension = ".cshtml";
+
+    public static bool IsRazorFilePath(this string filePath)
+    {
+        var comparison = PathUtilities.OSSpecificPathComparison;
+
+        return filePath.EndsWith(RazorExtension, comparison) ||
+               filePath.EndsWith(CSHtmlExtension, comparison);
+    }
+
     public static int? GetFirstNonWhitespaceOffset(this string line)
     {
         for (var i = 0; i < line.Length; i++)
